extracting parsing tests
