+++ Testing/parsing/tests.py
@@ -1,0 +1,3 @@
+from test_rule import TestRule
+
+TestRule.test_parser()
